Refactor classifier code organization and fix napari viewer usage of scale
