+++ napari_feature_visualization/dockwidgets.py
@@ -1,6 +1,6 @@
 from napari_plugin_engine import napari_hook_implementation
 from .feature_vis import feature_vis
-from .classifier import initialize_classifier, load_classifier
+from .classifier_widgets import initialize_classifier, load_classifier
 
 @napari_hook_implementation
 def napari_experimental_provide_dock_widget():
